refactor: move entities to folder, id to long
Move Entity and VersionedEntity into an Entities folder and namespace,
updating all referencing usings.

Widen the entity identifier from int to long, cascading through the
repository interfaces and EfRepository (GetById params and the Create,
CreateRange, Delete, DeleteRange return types) and the affected tests.

Refresh the README and architecture diagrams to match, and fix the docs
landing-page links to use the baseURL-prefixed paths.

+++ src/ArturRios.Data.Relational.Core/Interfaces/IAsyncReadOnlyRepository.cs
@@ -1,3 +1,4 @@
+using ArturRios.Data.Relational.Core.Entities;
 using ArturRios.Output;
 
 namespace ArturRios.Data.Relational.Core.Interfaces;
@@ -20,5 +21,5 @@ public interface IAsyncReadOnlyRepository<T> where T : Entity
     ///     Returns the entity with the given identifier, or a successful result with
     ///     <c>null</c> data when none matches.
     /// </summary>
-    Task<DataOutput<T?>> GetByIdAsync(int id, CancellationToken ct = default);
+    Task<DataOutput<T?>> GetByIdAsync(long id, CancellationToken ct = default);
 }
